ShortTerm/requirements.txt
# ShortTerm - 事件驱动短线策略依赖
# 数据通过 DataHub 获取

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 统计分析
scipy>=1.10.0

# 配置解析
pyyaml>=6.0
